[RGZ] Refactored tests to be more maintainable.
