+++ partner_multi_relation/models/res_partner_relation.py
@@ -2,7 +2,8 @@
 # © 2013-2016 Therp BV <http://therp.nl>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 """Store relations (connections) between partners."""
-from openerp import _, api, exceptions, fields, models
+from openerp import _, api, fields, models
+from openerp.exceptions import ValidationError
 
 
 class ResPartnerRelation(models.Model):
@@ -53,11 +54,11 @@ class ResPartnerRelation(models.Model):
     def _check_dates(self):
         """End date should not be before start date, if not filled
 
-        :raises exceptions.Warning: When constraint is violated
+        :raises ValidationError: When constraint is violated
         """
         if (self.date_start and self.date_end and
                 self.date_start > self.date_end):
-            raise exceptions.Warning(
+            raise ValidationError(
                 _('The starting date cannot be after the ending date.')
             )
 
@@ -66,7 +67,7 @@ class ResPartnerRelation(models.Model):
     def _check_partner_left(self):
         """Check left partner for required company or person
 
-        :raises exceptions.Warning: When constraint is violated
+        :raises ValidationError: When constraint is violated
         """
         self._check_partner("left")
 
@@ -75,7 +76,7 @@ class ResPartnerRelation(models.Model):
     def _check_partner_right(self):
         """Check right partner for required company or person
 
-        :raises exceptions.Warning: When constraint is violated
+        :raises ValidationError: When constraint is violated
         """
         self._check_partner("right")
 
@@ -84,20 +85,20 @@ class ResPartnerRelation(models.Model):
         """Check partner for required company or person, and for category
 
         :param str side: left or right
-        :raises exceptions.Warning: When constraint is violated
+        :raises ValidationError: When constraint is violated
         """
         assert side in ['left', 'right']
         ptype = getattr(self.type_id, "contact_type_%s" % side)
         partner = getattr(self, '%s_partner_id' % side)
         if ((ptype == 'c' and not partner.is_company) or
                 (ptype == 'p' and partner.is_company)):
-            raise exceptions.Warning(
+            raise ValidationError(
                 _('The %s partner is not applicable for this relation type.') %
                 side
             )
         category = getattr(self.type_id, "partner_category_%s" % side)
         if category and category.id not in partner.category_id.ids:
-            raise exceptions.Warning(
+            raise ValidationError(
                 _('The %s partner does not have category %s.') %
                 (side, category.name)
             )
@@ -107,11 +108,11 @@ class ResPartnerRelation(models.Model):
     def _check_not_with_self(self):
         """Not allowed to link partner to same partner
 
-        :raises exceptions.Warning: When constraint is violated
+        :raises ValidationError: When constraint is violated
         """
         if self.left_partner_id == self.right_partner_id:
             if not (self.type_id and self.type_id.allow_self):
-                raise exceptions.Warning(
+                raise ValidationError(
                     _('Partners cannot have a relation with themselves.')
                 )
 
@@ -127,7 +128,7 @@ class ResPartnerRelation(models.Model):
         """Forbid multiple active relations of the same type between the same
         partners
 
-        :raises exceptions.Warning: When constraint is violated
+        :raises ValidationError: When constraint is violated
         """
         # pylint: disable=no-member
         # pylint: disable=no-value-for-parameter
@@ -150,6 +151,6 @@ class ResPartnerRelation(models.Model):
                 ('date_start', '<=', self.date_end),
             ]
         if self.search(domain):
-            raise exceptions.Warning(
+            raise ValidationError(
                 _('There is already a similar relation with overlapping dates')
             )
