+++ partner_multi_relation/tests/__init__.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
 # Copyright 2016 Therp BV
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import test_partner_multi_relation
+from . import test_partner_relation_common
+from . import test_partner_relation
+from . import test_partner_relation_all
+from . import test_partner_search
